change condition of delete user when fail authenticate with email

+++ backend/accountapp/views.py
@@ -70,7 +70,7 @@ class ActivateAccountAPI(APIView):
             print('이메일 인증이 성공하였습니다.') # 검토후 삭제
             return Response({'message': '이메일 인증이 완료되었습니다.'})
         else:
-            if user is not None:
+            if user is not None and user.is_active==False:
                 user.delete()
             print('이메일 인증이 실패하였습니다.') # 검토후 삭제
             return Response({'message': '이메일 인증이 실패하였습니다.'})
